# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_import_dataset_api_v1_datasets_import_post": {
        "properties": {
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "dataset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Name"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "mapping": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_dataset_api_v1_datasets_import_post",
        "type": "object"
      },
...
      "CaseResponse": {
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "reference_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Answer"
          },
          "reference_chunks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReferenceChunkResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Chunks"
          }
        },
        "required": [
          "case_id",
          "question",
          "reference_answer",
          "reference_chunks"
        ],
        "title": "CaseResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportFailure": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ImportIssueResponse"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "error",
          "issues"
        ],
        "title": "ImportFailure",
        "type": "object"
      },
      "ImportIssueResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "line",
          "field",
          "code",
          "message"
        ],
        "title": "ImportIssueResponse",
        "type": "object"
      },
...
      "ReferenceChunkResponse": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "document_id"
        ],
        "title": "ReferenceChunkResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionDetail": {
        "properties": {
          "case_count": {
            "title": "Case Count",
            "type": "integer"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseResponse"
            },
            "title": "Cases",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "source_collection_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Collection Ids",
            "type": "array"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "dataset_id",
          "version",
          "case_count",
          "source_filename",
          "created_at",
          "total",
          "offset",
          "limit",
          "cases"
        ],
        "title": "VersionDetail",
        "type": "object"
      },
      "VersionSummary": {
        "properties": {
          "case_count": {
            "title": "Case Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source_collection_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Collection Ids",
            "type": "array"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "dataset_id",
          "version",
          "case_count",
          "source_filename",
          "created_at"
        ],
        "title": "VersionSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ragEva API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datasets/import": {
      "post": {
        "operationId": "import_dataset_api_v1_datasets_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_dataset_api_v1_datasets_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportFailure"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Import Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/versions": {
      "get": {
        "operationId": "list_versions_api_v1_datasets__dataset_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionSummary"
                  },
                  "title": "Response List Versions Api V1 Datasets  Dataset Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Versions",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/versions/{version}": {
      "get": {
        "operationId": "get_version_api_v1_datasets__dataset_id__versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Version",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/versions/{version}/cases/{case_id}": {
      "get": {
        "operationId": "get_case_api_v1_datasets__dataset_id__versions__version__cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Case",
        "tags": [
          "datasets"
        ]
      }
    },
...
  }
}